Removed class Fast_dtft, and added a function for fft in main
