--- scripts/fast_dtft.py
@@ -1,56 +1,0 @@
-#!/usr/bin/env python
-
-import numpy as np
-import cfg 
-
-from math import atan2
-from scipy import fft, ifft, fftpack
-
-
-class FastDTFT:
-  def __init__(self):
-    self.M = cfg.M
-    self.Fs = cfg.Fs
-
-  # def zero_pad(self, signal, M=None):
-  #   if M is None:
-  #     M = self.M
-
-  #   # zero-padding such that the system can take M-point FFT
-  #   for i in range(M - len(signal)):
-  #     signal.append(0)
-  #   return signal
-
-  def fast_dtft(self, signal):
-    # M number of fft-points
-    # if len(signal) < self.M:
-    #   signal = self.zero_pad(signal)       
-    Fx = fft.fft(signal, self.M)
-    Ff = fftpack.fftfreq(self.M, 1 / self.Fs)
-
-    Fx = fftpack.fftshift(Fx)
-    Ff = fftpack.fftshift(Ff)
-    
-    return Fx, Ff
-
-  def magnitude(self, signal):
-    # Return || signal ||_{2}
-    magnitude = [0 for i in range(len(signal))]
-    
-    for i in range(len(signal)):
-      magnitude[i] = abs(signal[i])
-
-    return magnitude
-
-  def phase(self, signal):
-    # Calculates the phase of a given signal
-    # Each time-point must later be merged with the equivalent frequency to
-    # get the optimal response
-    phase = [0 for i in range(len(signal))] 
-
-    for i in range(len(signal)):
-      phase[i] = -atan2(signal[i].imag, signal[i].real)
-    
-    return phase
-
-
